pyproject.toml
[project]
name = "joystick-mouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogui>=0.9.54",
    "pyserial>=3.5",
    "pyyaml>=6.0.2",
]
